# docs/swagger.yaml
basePath: /
definitions:
...
  models.CreateRole:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
  models.Permission:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  models.Role:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  models.Role_Permission_Payload:
    properties:
      permissions:
        items:
          type: integer
        type: array
    type: object
...
  models.User_Role_Payload:
    properties:
      users:
        items:
          type: string
        type: array
    type: object
...
info:
  contact:
    name: API Support
  description: This is an RBAC based full fledge API serice for managing users, roles
    and permissions in the system
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Truvest Identity Management Service APIs
  version: "1.0"
paths:
...
  /roles:
    get:
      consumes:
      - application/json
      description: Get all roles in the system. In order to access this API, someone
        must have "ROLES_VIEW" Permission tagged to its role.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Role'
      security:
      - ApiKeyAuth: []
      summary: Get all Roles in the system
      tags:
      - Roles
    post:
      consumes:
      - application/json
      description: Create a role in the system. User must have "MANAGE_ROLES" permission
        tagged to its role in order to use this API.
      parameters:
      - description: Create Role
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.CreateRole'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Role'
      security:
      - ApiKeyAuth: []
      summary: Create a role in the system
      tags:
      - Role
  /roles/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a role by id in the system. In order to access this API,
        someone must have "MANAGE_ROLES" Permission tagged to its role.
      parameters:
      - description: id of the Role
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: Delete a role by id in the system
      tags:
      - Role
    get:
      consumes:
      - application/json
      description: Get a role by id in the system. In order to access this API, someone
        must have "ROLES_VIEW" Permission tagged to its role.
      parameters:
      - description: ID of the role to get
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Role'
      security:
      - ApiKeyAuth: []
      summary: Get a Role by id in the system
      tags:
      - Role
    put:
      consumes:
      - application/json
      description: Update a role by id in the system. In order to access this API,
        someone must have "MANAGE_ROLES" Permission tagged to its role.
      parameters:
      - description: id of the Role
        in: path
        name: id
        required: true
        type: integer
      - description: Create Role
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.CreateRole'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Role'
      security:
      - ApiKeyAuth: []
      summary: Update a role by id in the system
      tags:
      - Role
  /roles/{id}/permissions:
    post:
      consumes:
      - application/json
      description: Add permissions to role by id in the system. In order to access
        this API, someone must have "MANAGE_PERMISSION" and "PERMISSION_ASSIGN_TO_ROLE"
        Permission tagged to its role.
      parameters:
      - description: id of the Role
        in: path
        name: id
        required: true
        type: integer
      - description: Role Permission Payload
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/models.Role_Permission_Payload'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: Add permissions to role by id in the system
      tags:
      - Role
  /roles/{id}/users:
    post:
      consumes:
      - application/json
      description: Add users to role by id in the system. In order to access this
        API, someone must have "USERS_ASSIGN_TO_ROLE" Permission tagged to its role.
      parameters:
      - description: id of the Role
        in: path
        name: id
        required: true
        type: integer
      - description: User Role Payload
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/models.User_Role_Payload'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: Add users to role by id in the system
      tags:
      - Role
  /roles/{id1}/permissions/{id2}:
    delete:
      consumes:
      - application/json
      description: Delete permissions from role by id in the system.
      parameters:
      - description: id of the Role
        in: path
        name: id1
        required: true
        type: integer
      - description: id of the Permission
        in: path
        name: id2
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: Delete permissions from role by id in the system
      tags:
      - Role
  /roles/{id1}/users/{id2}:
    delete:
      consumes:
      - application/json
      description: Delete users from role by id in the system. In order to access
        this API, someone must have
      parameters:
      - description: id of the Role
        in: path
        name: id1
        required: true
        type: integer
      - description: id of the User
        in: path
        name: id2
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: Delete users from role by id in the system
      tags:
      - Role
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"